flask_app/APP/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.5.0
cryptography==2.4.2
ez-setup==0.9
Flask==1.0.2
Flask-WTF==0.14.2
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
multitasking==0.0.7
mysql-connector==2.1.6
mysql-connector-python==8.0.13
mysqlclient==1.3.14
numpy==1.15.4
pandas==0.23.4
paramiko==2.4.2
passlib==1.7.1
protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.4
pycparser==2.19
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
Werkzeug==0.14.1
WTForms==2.2.1
